+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Versions
 
 ## x.x.x
+* Add APIs for Selective Init. For more info, check out our [docs](https://dash.applovin.com/documentation/mediation/react-native/getting-started/advanced-settings#selective-init).
 * Fix TypeScript compilation errors. (https://github.com/AppLovin/AppLovin-MAX-React-Native/issues/271)
 ## 6.0.2
 * Depend on Android SDK 12.0.0 and iOS SDK 12.0.0.